support both aws-sdk 1 and 2
