+++ lib/coverband.rb
@@ -8,7 +8,7 @@ require 'coverband/configuration'
 require 'coverband/adapters/base'
 require 'coverband/adapters/redis_store'
 require 'coverband/adapters/file_store'
-require 'coverband/utils/s3_report_writer'
+require 'coverband/utils/s3_report'
 require 'coverband/collectors/coverage'
 require 'coverband/reporters/base'
 require 'coverband/reporters/simple_cov_report'
