+++ lib/coverband/version.rb
@@ -1,5 +1,5 @@
 # frozen_string_literal: true
 
 module Coverband
-  VERSION = '3.0.0'
+  VERSION = '3.0.1.alpha'
 end
